+++ CHANGELOG.md
@@ -69,6 +69,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   string being freed is an HTML-like string.
 - **Breaking**: The `agcanon` function has been removed.
 - **Breaking**: The `Agiddisc_t.alloc` member has been removed.
+- **Breaking**: The `cflag` parameter to `agidsubg` has been removed.
 
 ### Fixed